+++ src/Container/Container.js
@@ -4,24 +4,7 @@ import PropTypes from 'prop-types';
 import { useStripes } from '@folio/stripes/core';
 
 import View from '../View';
-
-const SELECTED_STATUS = 'status.selected';
-const UNSELECTED_STATUS = 'status.unselected';
-
-export const filterApplications = (applications, checkedAppIdsMap, filter, query) => {
-  if (filter === SELECTED_STATUS) {
-    return applications.filter(app => app.id in checkedAppIdsMap);
-  }
-
-  if (filter === UNSELECTED_STATUS) {
-    return applications.filter(app => !(app.id in checkedAppIdsMap));
-  }
-
-  if (query) {
-    return applications.filter(app => app.id.includes(query));
-  }
-  return applications;
-};
+import { filterApplications } from '../Utils';
 
 export default function Container({
   onClose,
@@ -60,9 +43,4 @@ Container.propTypes = {
   onClose: PropTypes.func.isRequired,
   onSave: PropTypes.func.isRequired,
   checkedAppIdsMap: PropTypes.object
-};
-
-module.exports = {
-  Container,
-  filterApplications
 };
